requirements.txt
arxiv>=2.1.0
openai>=1.30.0
google-genai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pyzotero>=1.5.0
pymongo>=4.6.0
